# cmd/docs/swagger.yaml
definitions:
...
  aqary_admin_internal_domain_requests_user.CreateRolePermissionRequest:
    properties:
      permissions_id:
        items:
          type: integer
        type: array
      roles_id:
        type: integer
      sub_section_id:
        items:
          type: integer
        type: array
    required:
    - permissions_id
    - roles_id
    type: object
...
  aqary_admin_internal_domain_requests_user.DeleteOneRolePermissionRequest:
    properties:
      permissions_id:
        type: integer
    required:
    - permissions_id
    type: object
...
  aqary_admin_internal_domain_requests_user.UpdateRolePermissionRequest:
    properties:
      permissions_id:
        items:
          type: integer
        type: array
      sub_section_id:
        items:
          type: integer
        type: array
    required:
    - permissions_id
    type: object
...
  aqary_admin_internal_domain_responses_user.CustomAllQuaternarySubSectionPermission:
    properties:
      id:
        type: integer
      indicator:
        type: integer
      label:
        type: string
      permission_id:
        type: integer
      primary_id:
        type: integer
      quinary_permission:
        items:
          $ref: '#/definitions/aqary_admin_internal_domain_responses_user.CustomAllQuinarySubSectionPermission'
        type: array
      secondary_id:
        type: integer
      sub_label:
        type: string
      sub_section_button_action:
        type: string
      sub_section_button_id:
        type: integer
      tertiary_id:
        type: integer
    type: object
  aqary_admin_internal_domain_responses_user.CustomAllQuinarySubSectionPermission:
    properties:
      id:
        type: integer
      indicator:
        type: integer
      label:
        type: string
      permission_id:
        type: integer
      primary_id:
        type: integer
      quaternary_id:
        type: integer
      secondary_id:
        type: integer
      sub_label:
        type: string
      sub_section_button_action:
        type: string
      sub_section_button_id:
        type: integer
      tertiary_id:
        type: integer
    type: object
  aqary_admin_internal_domain_responses_user.CustomAllSecondarySubSectionPermission:
    properties:
      id:
        type: integer
      indicator:
        type: integer
      label:
        type: string
      permission_id:
        type: integer
      primary_id:
        type: integer
      quaternary_permission:
        items:
          $ref: '#/definitions/aqary_admin_internal_domain_responses_user.CustomAllQuaternarySubSectionPermission'
        type: array
      secondary_id:
        type: integer
      sub_label:
        type: string
      sub_section_button_action:
        type: string
      sub_section_button_id:
        type: integer
    type: object
  aqary_admin_internal_domain_responses_user.CustomAlllPermission:
    properties:
      id:
        type: integer
      indicator:
        type: integer
      label:
        type: string
      secondary_permission:
        items:
          $ref: '#/definitions/aqary_admin_internal_domain_responses_user.CustomSubSectionSecondaryPermission'
        type: array
      sub_label:
        type: string
    type: object
  aqary_admin_internal_domain_responses_user.CustomRolePermission:
    properties:
      id:
        type: integer
      label:
        type: string
      section_permissions:
        items:
          $ref: '#/definitions/aqary_admin_internal_domain_responses_user.CustomSectionPermission'
        type: array
    type: object
  aqary_admin_internal_domain_responses_user.CustomSectionPermission:
    properties:
      id:
        type: integer
      indicator:
        type: integer
      label:
        type: string
      permissions:
        items:
          $ref: '#/definitions/aqary_admin_internal_domain_responses_user.CustomAlllPermission'
        type: array
      sub_label:
        type: string
    type: object
  aqary_admin_internal_domain_responses_user.CustomSubSectionSecondaryPermission:
    properties:
      id:
        type: integer
      indicator:
        type: integer
      label:
        type: string
      permission_id:
        type: integer
      primary_id:
        type: integer
      sub_label:
        type: string
      sub_section_button_action:
        type: string
      sub_section_button_id:
        type: integer
      tertiary_permission:
        items:
          $ref: '#/definitions/aqary_admin_internal_domain_responses_user.CustomAllSecondarySubSectionPermission'
        type: array
    type: object
...
  aqary_admin_internal_domain_sqlc_sqlc.RolesPermission:
    properties:
      created_at:
        type: string
      id:
        type: integer
      permissions_id:
        items:
          type: integer
        type: array
      roles_id:
        type: integer
      sub_section_permission:
        items:
          type: integer
        type: array
      updated_at:
        type: string
    type: object
...
  aqary_admin_pkg_utils.ErrResponseSwagger:
    properties:
      error:
        example: any error message
        type: string
    type: object
  aqary_admin_pkg_utils_exceptions.ErrorCode:
    enum:
    - 00SWR
    - 000BR
    - 00UNA
    - 00ILC
    - 000PD
    - 00UNF
    - CGEAU
    - CNVEC
    - 000AE
    - 00UCV
    - 00UAE
    - 00EAE
    - 00EAE
    - 0PTAE
    - 0DTAP
    - 00IAT
    - 0ATNF
    - 0ACNU
    - 00DAT
    - 0MUNF
    - 0PUNF
    - 0DUNF
    type: string
    x-enum-varnames:
    - SomethingWentWrongErrorCode
    - BadRequestErrorCode
    - UserNotAuthorizedErrorCode
    - InvalidCredentialsErrorCode
    - PermissionDeniedErrorCode
    - NoDataFoundErrorCode
    - CannotGiveEmailAndUsernameErrorCode
    - CompanyNotVerifiedErrorCode
    - AlreadyExistCode
    - UniqueConstraintViolationErrorCode
    - UsernameAlreadyExistsErrorCode
    - EmailAlreadyExistsErrorCode
    - EmailAlreadyExistErrorCode
    - DuplicatePlanTitleErrorCode
    - DuplicateDocumentTypeErrorCode
    - InvalidAuctionTypeErrorCode
    - ActivityTypeNotFountErrorCode
    - AuctionCanNotUpdateErrorCode
    - DuplicateAuctionTitleErrorCode
    - MediaNotFoundErorrCode
    - PlanDocumentNotFoundErrorCode
    - DocumentNotFoundErrorCode
  aqary_admin_pkg_utils_exceptions.ErrorMessage:
    enum:
    - something went wrong
    - bad request
    - user not authorized,please login
    - invalid login credentials
    - permission denied for this action
    - no data found
    - cannot give email and username
    - your company is not verified
    - already exist
    - a unique constraint was violated
    - username already exists
    - email already exists
    - one account already exist with this email, please try with another one
    - One auction already exist with the same auction title, Please create a diffrent
      title
    - A Plan with this plan title is already exist
    - A Document with this document type is already exist
    - Invalid request, please pass a valid auctionType
    - Invalid request, please pass a valid auctionId
    - Invalid auctionType value, it should be ID of Auction Type
    - Invalid request, please pass a valid activityType
    - Invalid activity type, no activity found for provided activity type
    - Auction can update only when it is in draft state
    - This media url is not exist in Media
    - This plan url is not exist in Plan
    - This document is not exist in Document
    type: string
    x-enum-varnames:
    - somethingWentWrong
    - badRequestMessage
    - userNotAuthorizedMessage
    - invalidCredentialsMessage
    - permissionDeniedErrormessage
    - noDataFoundErrormessage
    - cannotGiveEmailAndUsernameErrorCode
    - companyNotVerifiedErrorCode
    - alreadyExistCode
    - uniqueConstraintViolationMessage
    - usernameAlreadyExistsMessage
    - emailAlreadyExistsMessage
    - emailAlreadyExistErrorMessage
    - DuplicateAuctionTitleErrorMessage
    - duplicatePlanTitleErrorMessage
    - duplicateDocumentTypeErrorMessage
    - AuctionTypeRequiredErrorMessage
    - AuctionIdRequiredForListErrorMessage
    - InvalidAuctionTypeErrorMessage
    - ActivityTypeRequiredErrorMessage
    - ActivityTypeNotFountErrorMessage
    - AuctionCanNotUpdateErrorMessage
    - MediaNotFoundErorrMessage
    - PlanDocumentNotFoundErrorMessage
    - DocumentNotFoundErrorMessage
  aqary_admin_pkg_utils_exceptions.Exception:
    properties:
      errorCode:
        $ref: '#/definitions/aqary_admin_pkg_utils_exceptions.ErrorCode'
      errorMessage:
        $ref: '#/definitions/aqary_admin_pkg_utils_exceptions.ErrorMessage'
      httpCode:
        type: integer
    type: object
...
info:
  contact: {}
  description: Testing Swagger APIs.
  title: Aqary  APIs
  version: "1.0"
paths:
...
  /api/user/createRolePermission:
    post:
      consumes:
      - application/json
      description: Create a new role permission
      parameters:
      - description: Role Permission details
        in: body
        name: rolePermission
        required: true
        schema:
          $ref: '#/definitions/aqary_admin_internal_domain_requests_user.CreateRolePermissionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/aqary_admin_internal_domain_sqlc_sqlc.RolesPermission'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/aqary_admin_pkg_utils.ErrResponseSwagger'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/aqary_admin_pkg_utils.ErrResponseSwagger'
      summary: Create a new role permission
      tags:
      - role-permissions
...
  /api/user/deleteAllRolePermission/{id}:
    delete:
      consumes:
      - application/json
      description: Delete all permissions for an existing role
      parameters:
      - description: Role ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/aqary_admin_pkg_utils.ErrResponseSwagger'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/aqary_admin_pkg_utils.ErrResponseSwagger'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/aqary_admin_pkg_utils.ErrResponseSwagger'
      summary: Delete all permissions for a role
      tags:
      - role-permissions
...
  /api/user/deleteOneRolePermission/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a single permission from an existing role
      parameters:
      - description: Role Permission ID
        in: path
        name: id
        required: true
        type: integer
      - description: Permission to delete
        in: body
        name: permission
        required: true
        schema:
          $ref: '#/definitions/aqary_admin_internal_domain_requests_user.DeleteOneRolePermissionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/aqary_admin_internal_domain_sqlc_sqlc.RolesPermission'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/aqary_admin_pkg_utils.ErrResponseSwagger'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/aqary_admin_pkg_utils.ErrResponseSwagger'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/aqary_admin_pkg_utils.ErrResponseSwagger'
      summary: Delete a single permission from a role
      tags:
      - role-permissions
...
  /api/user/getAllPermissionWithoutPagination:
    get:
      consumes:
      - application/json
      description: Get all permissions without pagination
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/aqary_admin_internal_domain_responses_user.CustomSectionPermission'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/aqary_admin_pkg_utils_exceptions.Exception'
      summary: Get all permissions without pagination
      tags:
      - permissions
...
  /api/user/getAllRolePermission:
    get:
      consumes:
      - application/json
      description: Get all role permissions with pagination
      parameters:
      - description: Page number
        in: query
        name: page_no
        required: true
        type: integer
      - description: Page size
        in: query
        name: page_size
        required: true
        type: integer
      - description: Search term
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/aqary_admin_internal_domain_responses_user.CustomRolePermission'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/aqary_admin_pkg_utils.ErrResponseSwagger'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/aqary_admin_pkg_utils.ErrResponseSwagger'
      summary: Get all role permissions
      tags:
      - role-permissions
  /api/user/getAllRolePermissionWithoutPagination/all:
    get:
      consumes:
      - application/json
      description: Get all role permissions without pagination
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/aqary_admin_internal_domain_responses_user.CustomRolePermission'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/aqary_admin_pkg_utils.ErrResponseSwagger'
      summary: Get all role permissions without pagination
      tags:
      - role-permissions
...
  /api/user/updateRolePermission/{id}:
    put:
      consumes:
      - application/json
      description: Update an existing role permission
      parameters:
      - description: Role Permission ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated role permission details
        in: body
        name: rolePermission
        required: true
        schema:
          $ref: '#/definitions/aqary_admin_internal_domain_requests_user.UpdateRolePermissionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/aqary_admin_internal_domain_sqlc_sqlc.RolesPermission'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/aqary_admin_pkg_utils.ErrResponseSwagger'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/aqary_admin_pkg_utils.ErrResponseSwagger'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/aqary_admin_pkg_utils.ErrResponseSwagger'
      summary: Update a role permission
      tags:
      - role-permissions
...
securityDefinitions:
  bearerToken:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"